only give bootcamp modification permission to the bootcamp owner and an admin

+++ controllers/bootcamps.js
@@ -34,14 +34,22 @@ export const createBootcamp = asyncHandler(async (req, res, next) => {
 })
 
 export const updateBootcamp = asyncHandler(async (req, res, next) => {
-  const bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, {
-    new: true,
-    runValidators: true,
-  })
+  let bootcamp = await Bootcamp.findById(req.params.id)
 
   if (!bootcamp) {
     return new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
   }
+
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(`User ${req.params.id} is not authorized to update this bootcamp`, 401)
+    )
+  }
+
+  bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, {
+    new: true,
+    runValidators: true,
+  })
 
   res.status(200).json({ success: true, data: bootcamp })
 })
@@ -51,6 +59,12 @@ export const deleteBootcamp = asyncHandler(async (req, res, next) => {
 
   if (!bootcamp) {
     return new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
+  }
+
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(`User ${req.params.id} is not authorized to delete this bootcamp`, 401)
+    )
   }
 
   await bootcamp.deleteOne()
@@ -63,6 +77,12 @@ export const bootcampPhotoUpload = asyncHandler(async (req, res, next) => {
 
   if (!bootcamp) {
     return new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
+  }
+
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(`User ${req.params.id} is not authorized to update this bootcamp`, 401)
+    )
   }
 
   if (!req.files) {
